move chrono support to own module
