+++ macros/src/types/tuple.rs
@@ -3,12 +3,11 @@ use quote::quote;
 use syn::{Field, FieldsUnnamed, Generics, Result};
 
 use crate::{
-    attr::{FieldAttr, },
+    attr::{FieldAttr, StructAttr},
     deps::Dependencies,
     types::generics::{format_generics, format_type},
     DerivedTS,
 };
-use crate::attr::StructAttr;
 
 pub(crate) fn tuple(
     attr: &StructAttr,
