+++ ts-rs/tests/chrono.rs
@@ -1,13 +1,21 @@
-use chrono::{Date, DateTime, Duration, FixedOffset, Local, NaiveDate, NaiveDateTime, NaiveTime, Utc};
+use chrono::{
+    Date, DateTime, Duration, FixedOffset, Local, NaiveDate, NaiveDateTime, NaiveTime, Utc,
+};
 use ts_rs::TS;
 
 #[test]
 fn chrono() {
     #[derive(TS)]
+    #[allow(dead_code)]
     struct Chrono {
         date: (NaiveDate, Date<Utc>, Date<Local>, Date<FixedOffset>),
         time: NaiveTime,
-        date_time: (NaiveDateTime, DateTime<Utc>, DateTime<Local>, DateTime<FixedOffset>),
+        date_time: (
+            NaiveDateTime,
+            DateTime<Utc>,
+            DateTime<Local>,
+            DateTime<FixedOffset>,
+        ),
         duration: Duration,
     }
