+++ ts-rs/tests/union_with_internal_tag.rs
@@ -1,4 +1,4 @@
-#![allow(dead_code)]
+#![allow(dead_code, clippy::blacklisted_name)]
 
 use serde::Serialize;
 use ts_rs::TS;
